Hide title in ClipboardWindow
