+++ App/Keyboard Handling/View/ClipboardWindow.swift
@@ -1,0 +1,24 @@
+import AppKit
+import ClipNinja
+
+class ClipboardWindow: KeyboardEventWindow {
+
+    override var frameAutosaveName: NSWindow.FrameAutosaveName {
+        "ClipboardWindow"
+    }
+
+    override init(
+        keyboardController: KeyboardController,
+        contentRect: NSRect,
+        styleMask style: NSWindow.StyleMask,
+        backing backingStoreType: NSWindow.BackingStoreType,
+        defer flag: Bool
+    ) {
+        super.init(keyboardController: keyboardController, contentRect: contentRect, styleMask: style, backing: backingStoreType, defer: flag)
+        titlebarAppearsTransparent = true
+        isOpaque = false
+        hasShadow = false
+        backgroundColor = NSColor(Colors.factory.backgroundColor)
+        collectionBehavior = .moveToActiveSpace
+    }
+}
